# src/routes/swagger.json
{
	"components": {
...
		"schemas": {
			"AcceptableDbms": {
				"enum": [
					"mysql",
					"oracle",
					"mariadb",
					"postgres",
					"cubrid"
				],
				"type": "string"
			},
			"UserInterface": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"token",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"$ref": "#/components/schemas/UserInterface"
			},
			"Service": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"method": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"entityName": {
						"type": "string"
					},
					"tableName": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"stage": {
						"$ref": "#/components/schemas/Stage"
					},
					"stageId": {
						"type": "number",
						"format": "double"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/ServiceColumn"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"method",
					"description",
					"entityName",
					"tableName",
					"status",
					"user",
					"application",
					"stage",
					"stageId",
					"meta",
					"columns",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Application": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nameSpace": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"trafficConfigs": {
						"items": {
							"$ref": "#/components/schemas/TrafficConfig"
						},
						"type": "array"
					},
					"stages": {
						"items": {
							"$ref": "#/components/schemas/Stage"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"nameSpace",
					"title",
					"description",
					"userId",
					"user",
					"services",
					"trafficConfigs",
					"stages",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrafficConfig": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"type": {
						"type": "string"
					},
					"maxCount": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"application",
					"type",
					"maxCount",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Stage": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"application": {
						"$ref": "#/components/schemas/Application"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"application",
					"services",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Meta": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"dataType": {
						"type": "string"
					},
					"originalFileName": {
						"type": "string"
					},
					"remoteFilePath": {
						"type": "string"
					},
					"filePath": {
						"type": "string"
					},
					"encoding": {
						"type": "string"
					},
					"extension": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "string"
					},
					"db": {
						"type": "string"
					},
					"dbUser": {
						"type": "string"
					},
					"pwd": {
						"type": "string"
					},
					"table": {
						"type": "string"
					},
					"dbms": {
						"$ref": "#/components/schemas/AcceptableDbms"
					},
					"rowCounts": {
						"type": "number",
						"format": "double"
					},
					"skip": {
						"type": "number",
						"format": "double"
					},
					"sheet": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/MetaColumn"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"dataType",
					"originalFileName",
					"remoteFilePath",
					"filePath",
					"encoding",
					"extension",
					"host",
					"port",
					"db",
					"dbUser",
					"pwd",
					"table",
					"dbms",
					"rowCounts",
					"skip",
					"sheet",
					"isActive",
					"user",
					"service",
					"columns",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceColumn": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hidden": {
						"type": "boolean"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"columnName",
					"type",
					"hidden",
					"service",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptableType": {
				"enum": [
					"bigint",
					"int",
					"double",
					"bit",
					"date",
					"datetime",
					"time",
					"varchar",
					"text",
					"longtext",
					"boolean"
				],
				"type": "string"
			},
			"MetaColumn": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"originalColumnName": {
						"type": "string"
					},
					"columnName": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/AcceptableType"
					},
					"originalType": {
						"$ref": "#/components/schemas/AcceptableType"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"isHidden": {
						"type": "boolean"
					},
					"isSearchable": {
						"type": "boolean"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"params": {
						"items": {
							"$ref": "#/components/schemas/MetaParam"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"originalColumnName",
					"columnName",
					"type",
					"originalType",
					"size",
					"order",
					"isHidden",
					"isSearchable",
					"meta",
					"params",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParamOperatorType": {
				"enum": [
					"lt",
					"lte",
					"gt",
					"gte",
					"like",
					"eq",
					"neq"
				],
				"type": "string"
			},
			"MetaParam": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"metaColumn": {
						"$ref": "#/components/schemas/MetaColumn"
					},
					"operator": {
						"$ref": "#/components/schemas/ParamOperatorType"
					},
					"description": {
						"type": "string"
					},
					"isRequired": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"metaColumn",
					"operator",
					"description",
					"isRequired",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApplicationParams": {
				"properties": {
					"nameSpace": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"nameSpace",
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StageParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetaColumnSaveParams": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"columnName": {
						"type": "string"
					},
					"isHidden": {
						"type": "boolean"
					},
					"isSearchable": {
						"type": "boolean"
					},
					"size": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							},
							{
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							}
						]
					},
					"type": {
						"$ref": "#/components/schemas/AcceptableType"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"originalColumnName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"columnName",
					"isHidden",
					"isSearchable",
					"type",
					"order",
					"originalColumnName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetaSaveParams": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"dataType": {
						"type": "string"
					},
					"columns": {
						"items": {
							"$ref": "#/components/schemas/MetaColumnSaveParams"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"dataType",
					"columns"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceSaveParams": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"method": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"entityName": {
						"type": "string"
					},
					"meta": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/MetaSaveParams"
							},
							{
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							}
						]
					}
				},
				"required": [
					"id",
					"title",
					"method",
					"description",
					"entityName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetaParamSaveParams": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"isRequired": {
						"type": "boolean"
					},
					"operator": {
						"$ref": "#/components/schemas/ParamOperatorType"
					}
				},
				"required": [
					"isRequired",
					"operator"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApplicationSaveParams": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nameSpace": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/ServiceSaveParams"
						},
						"type": "array"
					},
					"params": {
						"properties": {},
						"additionalProperties": {
							"items": {
								"$ref": "#/components/schemas/MetaParamSaveParams"
							},
							"type": "array"
						},
						"type": "object"
					}
				},
				"required": [
					"id",
					"nameSpace",
					"title",
					"description",
					"services",
					"params"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrafficConfigParam": {
				"properties": {
					"type": {
						"type": "string"
					},
					"maxCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type",
					"maxCount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"scopes": {}
			}
		}
	},
	"info": {
		"title": "api-gen",
		"version": "0.0.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/applications": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Put",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplicationParams"
							}
						}
					}
				}
			}
		},
		"/api/applications/{applicationId}/api-docs": {
			"get": {
				"operationId": "GetDocs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/applications/{applicationId}": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/applications/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/applications/{id}/stages": {
			"post": {
				"operationId": "PostStage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StageParams"
							}
						}
					}
				}
			}
		},
		"/api/applications/{id}/save": {
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplicationSaveParams"
							}
						}
					}
				}
			}
		},
		"/api/applications/{id}/traffic-configs": {
			"get": {
				"operationId": "GetTrafficConfigs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrafficConfig"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "PostTrafficConfigs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrafficConfig"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/TrafficConfigParam"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/rest/applications": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"#Applications"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "nameSpace",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rest/applications/{id}": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"#Applications"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/rest/services/{id}": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"#Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}